@@ -2,37 +2,27 @@
 /*                                                                      */
 /*  FileVariableFactory                                                 */
 /*                                                                      */
-/*  File can't be save in C8. So, different implementation to store     */
+/*  File can't be saved in C8. So, different implementation to store    */
 /*  files are possible, and the factory give access to the different    */
 /* formats                                                              */
 /* ******************************************************************** */
 package org.camunda.cherry.definition.filevariable;
 
-import io.camunda.zeebe.spring.client.exception.ZeebeBpmnError;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Service;
 
-import java.util.StringTokenizer;
+import java.util.Random;
 
+@Service
 public class FileVariableFactory {
-    public final static String BPMNERROR_INCORRECT_STORAGEDEFINITION = "INCORRECT_STORAGEDEFINITION";
-    Logger logger = LoggerFactory.getLogger(FileVariableFactory.class.getName());
 
-    /**
-     * Use the getInstance() method
-     */
-    private FileVariableFactory() {
-    }
+    private final Random rand = new Random();
 
-    public static FileVariableFactory getInstance() {
-        return new FileVariableFactory();
-    }
 
     /**
      * get the FileVariable object from the different information
      * StorageDefinition is a string like
      * "JSON" : the value is a JSON information, to be unSerialize
-     * "TEMPORARYDISK:<path>", and the value is a file name in this directory.
+     * "FOLDER:<path>", and the value is a file name in this directory.
      *
      * @param fileContainer information to access the file
      * @return a fileVariable
@@ -42,19 +32,14 @@ public class FileVariableFactory {
         if (fileContainer == null || fileContainer.content == null)
             return null;
 
-        FileVariableStorage storageType = extractStorageType(fileContainer.storageDefinition);
+        StorageDefinition storageDefinition = StorageDefinition.getFromString(fileContainer.storageDefinition);
 
-        switch (storageType) {
-            case FOLDER:
-                return new FileVariableFolder().fromFolder(fileContainer.storageDefinition, fileContainer.content.toString());
-
-            case TEMPFOLDER:
-                return new FileVariableTempFolder().fromTempFolder(fileContainer.content.toString());
-
-            case JSON:
-            default:
-                return new FileVariableJSON().fromJson(fileContainer.content.toString());
-        }
+        return switch (storageDefinition.type) {
+            case FOLDER -> new FileVariableFolder(this).fromFolder(storageDefinition, fileContainer.content.toString());
+            case CMIS -> new FileVariableCMIS(this).fromCmis(storageDefinition, fileContainer.content.toString());
+            case TEMPFOLDER -> new FileVariableTempFolder().fromTempFolder(fileContainer.content.toString());
+            case JSON -> new FileVariableJSON().fromJson(fileContainer.content.toString());
+        };
 
     }
 
@@ -66,56 +51,64 @@ public class FileVariableFactory {
      * @return the FileContainer (depends on the storageDefinition code)
      * @throws Exception if an error arrive
      */
-    public FileVariableReference setFileVariable(String storageDefinition, FileVariable fileVariableValue) throws Exception {
+    public FileVariableReference setFileVariable(StorageDefinition storageDefinition, FileVariable fileVariableValue) throws Exception {
         if (fileVariableValue == null)
             return null;
 
-        FileVariableStorage storageType = extractStorageType(storageDefinition);
-        if (storageType == null)
-            throw new ZeebeBpmnError(BPMNERROR_INCORRECT_STORAGEDEFINITION, "Error during storageDefinition[" + storageDefinition + "]");
 
         FileVariableReference fileContainer = new FileVariableReference();
-        fileContainer.storageDefinition = storageDefinition;
 
-        switch (storageType) {
+        fileContainer.storageDefinition = storageDefinition.encodeToString();
+
+        switch (storageDefinition.type) {
             case FOLDER:
-                fileContainer.content = new FileVariableFolder().toFolder(storageDefinition, fileVariableValue);
+                fileContainer.content = new FileVariableFolder(this).toFolder(storageDefinition, fileVariableValue);
                 break;
+
+            case CMIS:
+                fileContainer.content = new FileVariableCMIS(this).toCmis(storageDefinition, fileVariableValue);
+                break;
+
             case TEMPFOLDER:
                 fileContainer.content = new FileVariableTempFolder().toTempFolder(fileVariableValue);
                 break;
+
             case JSON:
-            default:
                 fileContainer.content = new FileVariableJSON().toJson(fileVariableValue);
+                break;
         }
         return fileContainer;
     }
 
-    /**
-     * Extract the storageType from the storageDefinition connection string
-     *
-     * @param storageDefinition the storage definition string
-     * @return the storage type
-     */
-    private FileVariableStorage extractStorageType(String storageDefinition) {
-        try {
-            StringTokenizer st = new StringTokenizer(storageDefinition, ":");
 
-            String storageTypeSt = st.hasMoreTokens() ? st.nextToken() : null;
-            return FileVariableStorage.valueOf(storageTypeSt);
-        } catch (Exception e) {
-            String message = "Can't decode storageDefinition [" + storageDefinition + "]. Format should be ["
-                    + FileVariableStorage.JSON
-                    + "|" + FileVariableStorage.TEMPFOLDER
-                    + "|" + FileVariableStorage.FOLDER + "]";
-            logger.error("Cherry.FileVariableFactory: " + message);
-            throw new ZeebeBpmnError(BPMNERROR_INCORRECT_STORAGEDEFINITION, message);
+    public boolean purgeFileVariable(FileVariableReference fileContainer) throws Exception {
+        if (fileContainer == null)
+            return true;
+
+        StorageDefinition storageDefinition = StorageDefinition.getFromString(fileContainer.storageDefinition);
+
+        switch (storageDefinition.type) {
+            case FOLDER:
+                return new FileVariableFolder(this).removeFile(storageDefinition, fileContainer.content.toString());
+
+            case CMIS:
+                new FileVariableCMIS(this).removeFile(storageDefinition, fileContainer.content.toString());
+                return true;
+
+            case TEMPFOLDER:
+                return new FileVariableTempFolder().removeFile(fileContainer.content.toString());
+
         }
+        return true;
     }
 
     /**
-     * Define how the file variable is stored.
-     * JSON: easy, but attention to large file
+     * Generate an uniq Identifier for class who search for one
+     *
+     * @return a uniq ID
      */
-    public enum FileVariableStorage {JSON, TEMPFOLDER, FOLDER}
+    public String generateUniqId() {
+        // get an uniq identifier
+        return "_" + System.currentTimeMillis() + "_" + rand.nextInt(10000);
+    }
 }
