@@ -7,16 +7,16 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class AbstractWorkerTest {
 
-  @Test
-  public void parseNumbersFromInputs() {
+    @Test
+    public void parseNumbersFromInputs() {
 
-    assertFalse(AbstractWorker.canParse(Integer.class, "foo"));
-    assertTrue(AbstractWorker.canParse(Integer.class, "01"));
-    assertTrue(AbstractWorker.canParse(Integer.class, "1001"));
-    assertTrue(AbstractWorker.canParse(Integer.class, 1001));
+        assertFalse(AbstractWorker.canParse(Integer.class, "foo"));
+        assertTrue(AbstractWorker.canParse(Integer.class, "01"));
+        assertTrue(AbstractWorker.canParse(Integer.class, "1001"));
+        assertTrue(AbstractWorker.canParse(Integer.class, 1001));
 
-    assertTrue(AbstractWorker.canParse(Long.class, "1001"));
-    assertTrue(AbstractWorker.canParse(Long.class, "1000000"));
-    assertTrue(AbstractWorker.canParse(Long.class, 1000000));
-  }
+        assertTrue(AbstractWorker.canParse(Long.class, "1001"));
+        assertTrue(AbstractWorker.canParse(Long.class, "1000000"));
+        assertTrue(AbstractWorker.canParse(Long.class, 1000000));
+    }
 }
